Updated for Sonar 5.2
Sonar 5.2 does not call decorators in incremental and preview builds anymore.
So the functionality has been moved to the PostJob

Fixes #11
