+++ pom.xml
@@ -32,7 +32,7 @@
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<sonar.buildVersion>4.5.1</sonar.buildVersion>
+		<sonar.buildVersion>5.2-SNAPSHOT</sonar.buildVersion>
 		<jdk.min.version>1.7</jdk.min.version>
 	</properties>
 
@@ -59,6 +59,12 @@
 			<groupId>org.codehaus.sonar</groupId>
 			<artifactId>sonar-testing-harness</artifactId>
 			<version>${sonar.buildVersion}</version>
+			<scope>test</scope>
+		</dependency>
+		<dependency>
+			<groupId>org.easytesting</groupId>
+			<artifactId>fest-assert</artifactId>
+			<version>1.4</version>
 			<scope>test</scope>
 		</dependency>
 		<dependency>
@@ -92,6 +98,12 @@
 			<groupId>com.intellij</groupId>
 			<artifactId>annotations</artifactId>
 			<version>12.0</version>
+		</dependency>
+		<dependency>
+			<groupId>org.slf4j</groupId>
+			<artifactId>slf4j-api</artifactId>
+			<version>1.7.10</version>
+			<scope>provided</scope>
 		</dependency>
 	</dependencies>
